[Gold III] Title: 나머지 합, Time: 136 ms, Memory: 5936 KB -BaekjoonHub

+++ 백준/Gold/10986. 나머지 합/나머지 합.cc
@@ -3,7 +3,7 @@
 using namespace std;
 
 int number[1000001];
-long long remainders[1001];
+long remainders[1001];
 
 int main(void)
 {
